Various fixes, tweaks, refactoring, renames,... Diet Genes should work again now. Removed comically misplaced comment Slime resource hopefully doesn't random reset not (thanks for that jank new 1.5 code there Ludeon...) FilterList better terminology. Still a lot of refactoring needed and hardcoded stuff from the old B&S framework left ot clean up though.
