+++ 1.5/Main/Source/BetterPrerequisites/SimpleCustomRaces/SimpleRaceCache.cs
@@ -134,7 +134,7 @@ namespace BigAndSmall
                 hediffsToRemove = pawn.health.hediffSet.hediffs.Where(h => 
                     hediffFilter.GetFilterResult(h.def).Denied() ||
                     // If a chemical dependency and not whitelisted, remove it.
-                    (h is Hediff_Addiction hd && banAddictions && hediffFilter.GetFilterResult(h.def).NeutralOrWorse())
+                    (h is Hediff_Addiction hd && banAddictions && hediffFilter.GetFilterResult(h.def).NotExplicitlyAllowed())
 
                 ).ToList();
             }
